Desktop/React+Fast APi/Backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
python-multipart
PyPDF2
openai
requests
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
slowapi
bleach